using Big integer

+++ app/src/main/java/com/example/fibonacciseries/NumbersRecyclerViewAdapter.kt
@@ -5,8 +5,9 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
+import java.math.BigInteger
 
-class NumbersRecyclerViewAdapter (var numbersList:List<Int>):RecyclerView.Adapter<NumbersViewHolder>() {
+class NumbersRecyclerViewAdapter (var numbersList:List<BigInteger>):RecyclerView.Adapter<NumbersViewHolder>() {
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): NumbersViewHolder {
         var itemView= LayoutInflater.from(parent.context)
             .inflate(R.layout.list_item_view,parent,false)
